refactor(desktop): colocate chat store with screen, mirroring cli layout
Adds apps/desktop/src/mainview/screens/chat/store.tsx, an
unstated createContainer that owns chat-feature-only state.
Mirrors the cli's per-screen store pattern
(apps/cli/src/screens/[name]/store.tsx) so navigating between
the two codebases follows the same conventions.

inputDeviceId moves from the cross-screen voice-store into the
new chat store — dictation is a chat-feature concern, not a
property of the global voice runtime (TTS/STT/health), which
stays in voice-store and remains mounted at the app root for
future non-chat screens that need it.

ChatStoreProvider is mounted inside ChatLayout so the chat
screen tree sees the container; chat-settings-sheet and
chat-panel now read inputDeviceId from useChatStore() instead
of useVoiceStore().
