+++ apps/desktop/src/mainview/screens/chat/chat-panel.tsx
@@ -29,6 +29,7 @@ import {
   MessageContent,
 } from "~/ui/message";
 import { Composer } from "./composer";
+import { useChatStore } from "./store";
 
 interface ChatPanelProps {
   model: string;
@@ -100,7 +101,8 @@ export function ChatPanel({
   // every render when the query is disabled (conversationId === null) and
   // `data` stays undefined.
   const persisted = useMemo<MessageRow[]>(() => data ?? [], [data]);
-  const { tts, speak, stopAudio, transcribe, inputDeviceId } = useVoiceStore();
+  const { tts, speak, stopAudio, transcribe } = useVoiceStore();
+  const { inputDeviceId } = useChatStore();
   const isSpeaking = tts.kind === "speaking";
 
   const createConv = useMutation({
